move src to app dir
