--- config/app.json
@@ -1,6 +1,0 @@
-{
-  "port": "3001",
-  "encoding": "utf8",
-  "proxyTarget": "http://local.com",
-  "resStatusKey": "code"
-}
