+++ config/env.js
@@ -1,0 +1,6 @@
+module.exports = {
+  port: "3001",
+  encoding: "utf8",
+  proxyTarget: "http://local.com",
+  resStatusKey: "code"
+};
