+++ config/pro.env.js
@@ -1,0 +1,5 @@
+const env = require("./env");
+
+module.exports = Object.assign(env, {
+  db: "mongodb://localhost/noobjs_dev"
+});
