+++ dev.js
@@ -1,4 +1,4 @@
 require("babel-register")({
   extensions: [".es6", ".es", ".jsx", ".js"]
 });
-require("./src/index");
+require("./app/index");
